- fixed 64bits compiling warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@6170 689a6050-402a-0410-94f2-e92a70836424

+++ modules/db_virtual/db_virtual.c
@@ -218,7 +218,7 @@ int add_set(char * name, char * mode){
 
 
     }else{
-        LM_DBG("alloc %p %i\n", global, sizeof(info_global_t));
+        LM_DBG("alloc %p %i\n", global, (int)sizeof(info_global_t));
         /* alloc global */
         LM_DBG("alloc %p\n", global);
         global = (info_global_t *) shm_malloc (1 * sizeof(info_global_t));
